Refactored code so that CAF does not act as the interface to the firmware specific commands, instead that is a newly created class in the Framework module, this could later be seperated into its on module if the need arises
